// Client/src/constants/category.enum.ts
export enum CategoryType {
  TrangChu = 0,
  LaptopMacbookSurface = 1,
  LaptopMayTinhXachTay = 2,
  LaptopAcer = 3,
  LaptopAcerAspire = 4,
  LaptopAcerSwift = 5,
  LaptopAcerSpin = 6,
  LaptopAcerTravelMate = 7,
  LaptopAsus = 8,
  LaptopAsusVivobook = 9,
  LaptopAsusZenbook = 10,
  LaptopAsusZenbookFlipS = 11,
  LaptopAsusZenbookFlip = 12,
  LaptopAsusZenbookDuo = 13,
  LaptopAsusExpertbook = 14,
  LaptopAsusXSeries = 15,
  LaptopAsusProArt = 16,
  LaptopDell = 17,
  LaptopDellInspiron = 18,
  LaptopDellVostro = 19,
  LaptopDellLatitude = 20,
  LaptopDellXps = 21,
  LaptopDellPrecision = 22,
  LaptopHp = 23,
  LaptopHpPavilion = 24,
  LaptopHpProbook = 25,
  LaptopHpElitebook = 26,
  LaptopHpEnvy = 27,
  LaptopHpSpectre = 28,
  LaptopHp14s15s = 29,
  LaptopHpPavilionX360 = 30,
  LaptopLenovo = 31,
  LaptopLenovoIdeapad = 32,
  LaptopLenovoThinkpad = 33,
  LaptopLenovoThinkbook = 34,
  LaptopLenovoYoga = 35,
  LaptopMsi = 36,
  LaptopMsiPrestige = 37,
  LaptopMsiSummit = 38,
  LaptopMsiModern = 39,
  LaptopMsiModern14 = 40,
  LaptopMsiModern15 = 41,
  Macbook = 42,
  LaptopAppleMacbookAir = 43,
  LaptopAppleMacbookPro13 = 44,
  LaptopLg = 45,
  LaptopLgGram14 = 46,
  LaptopLgGram15 = 47,
  LaptopLgGram16 = 48,
  LaptopLgGram17 = 49,
  MicrosoftSurface = 50,
  SurfacePro = 51,
  SurfaceGo = 52,
  SurfaceLaptop = 53,
  LaptopVaio = 54,
  LaptopVaioFe = 55,
  MayTinhBang = 56,
  DienThoaiIPhone = 57,
  MayDocSach = 58,
  DongHoThongMinh = 59
}